refactor(app): move internal/slack → app/slack
Slack client, handler, and xlsx helpers move into the app module.
Imports rewritten across cmd/, bot/, and their tests. app/go.mod
pins slack-go v0.20.0 + excelize v2.10.1 to match root (prevent
version drift when `cd app && go test` runs in CI).
