+++ cmd/agentdock/adapters.go
@@ -3,7 +3,7 @@ package main
 import (
 	"log/slog"
 
-	slackclient "github.com/Ivantseng123/agentdock/internal/slack"
+	slackclient "github.com/Ivantseng123/agentdock/app/slack"
 )
 
 // slackPosterAdapter wraps slackclient.Client to satisfy bot.SlackPoster interface.
