+++ go.mod
@@ -3,6 +3,7 @@ module github.com/Ivantseng123/agentdock
 go 1.25.0
 
 require (
+	github.com/Ivantseng123/agentdock/app v0.0.0-00010101000000-000000000000
 	github.com/Ivantseng123/agentdock/shared v0.0.0
 	github.com/Ivantseng123/agentdock/worker v0.0.0-00010101000000-000000000000
 	github.com/fsnotify/fsnotify v1.9.0
@@ -14,7 +15,6 @@ require (
 	github.com/slack-go/slack v0.20.0
 	github.com/spf13/cobra v1.10.2
 	github.com/spf13/pflag v1.0.9
-	github.com/xuri/excelize/v2 v2.10.1
 	golang.org/x/sync v0.20.0
 	golang.org/x/term v0.42.0
 	gopkg.in/yaml.v3 v3.0.1
@@ -48,6 +48,7 @@ require (
 	github.com/richardlehane/msoleps v1.0.6 // indirect
 	github.com/tiendc/go-deepcopy v1.7.2 // indirect
 	github.com/xuri/efp v0.0.1 // indirect
+	github.com/xuri/excelize/v2 v2.10.1 // indirect
 	github.com/xuri/nfp v0.0.2-0.20250530014748-2ddeb826f9a9 // indirect
 	go.uber.org/atomic v1.11.0 // indirect
 	go.yaml.in/yaml/v2 v2.4.2 // indirect
