+++ internal/bot/workflow_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/Ivantseng123/agentdock/internal/config"
 	"github.com/Ivantseng123/agentdock/shared/queue"
-	slackclient "github.com/Ivantseng123/agentdock/internal/slack"
+	slackclient "github.com/Ivantseng123/agentdock/app/slack"
 )
 
 // stubSlack implements the slackAPI interface for workflow tests.
